requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyMySQL
pandas
mysql-connector-python